some more staffpanel changes

+++ staff/views.py
@@ -595,26 +595,30 @@ class MatchDeclareWinner(View):
             return render(request, 'staff/permissiondenied.html')
         else:
             matchobj = Match.objects.get(pk=pk)
-            form = DeclareMatchWinnerPost(request.POST, instance=matchobj)
-            instance = form.instance
-            match = Match.objects.get(id=self.kwargs['pk'])
-            winner = Team.objects.get(id=form.data['winner'])
-            teams = list()
-            teams.append(match.hometeam)
-            teams.append(match.awayteam)
-            teams.remove(winner)
-            loser = teams[0]
-            instance.match = match
-            instance.winner = winner
-            instance.loser = loser
-            instance.completed = True
-            instance.save()
-            winner.num_matchwin += 1
-            loser.num_matchloss += 1
-            winner.save()
-            loser.save()
-            messages.success(request, "Winner declared")
-            return redirect('staff:matches_index')
+            if not matchobj.bye_2 and not matchobj.bye_1:
+                form = DeclareMatchWinnerPost(request.POST, instance=matchobj)
+                instance = form.instance
+                match = Match.objects.get(id=self.kwargs['pk'])
+                winner = Team.objects.get(id=form.data['winner'])
+                teams = list()
+                teams.append(match.hometeam)
+                teams.append(match.awayteam)
+                teams.remove(winner)
+                loser = teams[0]
+                instance.match = match
+                instance.winner = winner
+                instance.loser = loser
+                instance.completed = True
+                instance.save()
+                winner.num_matchwin += 1
+                loser.num_matchloss += 1
+                winner.save()
+                loser.save()
+                messages.success(request, "Winner declared")
+                return redirect('staff:matches_index')
+            else:
+                messages.error(request, 'Bye match, cannot set winner')
+                return redirect('staff:matches_index')
 
 
 def match_delete_winner(request, pk):
@@ -624,19 +628,23 @@ def match_delete_winner(request, pk):
         return render(request, 'staff/permissiondenied.html')
     else:
         match = Match.objects.get(pk=pk)
-        match.winner = None
-        match.completed = False
-        match.reported = False
-        match.team1reported = False
-        match.team2reported = False
-        match.team1reportedwinner = None
-        match.team2reportedwinner = None
-        match.disputed = False
-        match.save()
-        for i in MatchReport.objects.filter(match_id=pk):
-            i.delete()
-        messages.success(request, "Winner reset")
-        return redirect('staff:matches_index')
+        if not match.bye_1 and not match.bye_2:
+            match.winner = None
+            match.completed = False
+            match.reported = False
+            match.team1reported = False
+            match.team2reported = False
+            match.team1reportedwinner = None
+            match.team2reportedwinner = None
+            match.disputed = False
+            match.save()
+            for i in MatchReport.objects.filter(match_id=pk):
+                i.delete()
+            messages.success(request, "Winner reset")
+            return redirect('staff:matches_index')
+        else:
+            messages.error(request, 'Bye match, cannot change winner')
+            return redirect('staff:matches_index')
 
 
 def dispute_detail(request, pk):
